pyproject.toml
[project]
name = "agentgit"
version = "0.1.0"
description = "Convert agent transcripts to git repositories"
readme = "README.md"
authors = [
    { name = "Ben Tucker", email = "ben@btucker.net" }
]
requires-python = ">=3.10"
dependencies = [
    "pluggy>=1.3.0",
    "gitpython>=3.1.0",
    "click>=8.0.0",
]

[project.scripts]
agentgit = "agentgit.cli:main"

[project.optional-dependencies]
watch = [
    "watchdog>=3.0.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "watchdog>=3.0.0",
]

[project.entry-points."agentgit"]
claude_code = "agentgit.formats.claude_code:ClaudeCodePlugin"

[build-system]
requires = ["uv_build>=0.9.1,<0.10.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [
    "--cov=agentgit",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-fail-under=75",
]

[tool.coverage.run]
source = ["src/agentgit"]
branch = true
omit = ["*/tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]
show_missing = true
